*
git-svn-id: http://localhost:88/svn/uwow4/trunk/uwow2@258 3b753528-4e94-6941-993a-e4fe79dda6b7

+++ src/RealmServer/ObjectUpdater.cs
@@ -35,7 +35,7 @@ namespace Hazzik {
 			writer.Write((byte)GetUpdateType());
 			writer.WritePackGuid(_obj.Guid);
 			if(_isNew) {
-				writer.Write(_obj.TypeId);
+				writer.Write((byte)_obj.TypeId);
 				writer.Write((byte)(_obj != _to ? _obj.UpdateFlag : _obj.UpdateFlag | UpdateFlags.Self));
 				_obj.WriteCreateBlock(writer);
 				if(_obj.UpdateFlag.Has(UpdateFlags.HighGuid)) {
